Airflow/airdocker/requirements.txt
apache-airflow-providers-telegram
apache-airflow-providers-slack
apache-airflow-providers-http


# CAP modules
configparser
requests
pandas
openpyxl
psycopg2-binary
tqdm
logging
